% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_risk_stratification}
\alias{add_risk_stratification}
\title{Adicionar estratificação de risco}
\usage{
add_risk_stratification(data, config)
}
\description{
Adicionar estratificação de risco
}
\keyword{internal}
