% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{add_stage}
\alias{add_stage}
\title{Add a stage to a credit policy}
\usage{
add_stage(policy, stage)
}
\arguments{
\item{policy}{A \code{credit_policy} object.}

\item{stage}{A simulation stage created with \code{stage_cutoff}, \code{stage_rate}, etc.}
}
\value{
An updated \code{credit_policy} object.
}
\description{
Add a stage to a credit policy
}
