% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{analyze_tradeoffs_by_score}
\alias{analyze_tradeoffs_by_score}
\title{Analyze trade-offs by individual score}
\usage{
analyze_tradeoffs_by_score(results, cutoff_ranges)
}
\description{
Analyze trade-offs by individual score
}
\keyword{internal}
