% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{apply_keep_in_based_stage}
\alias{apply_keep_in_based_stage}
\title{Apply keep_in-based simulation stage}
\usage{
apply_keep_in_based_stage(
  data,
  stage_config,
  approval_col,
  stage_col,
  config,
  score_col
)
}
\description{
Apply keep_in-based simulation stage
}
\keyword{internal}
