% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{apply_random_removal_stage}
\alias{apply_random_removal_stage}
\title{Apply random removal simulation stage}
\usage{
apply_random_removal_stage(data, stage_config, approval_col, stage_col)
}
\description{
Apply random removal simulation stage
}
\keyword{internal}
