% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{apply_risk_aggravation}
\alias{apply_risk_aggravation}
\title{Aplicar agravamento por risco}
\usage{
apply_risk_aggravation(data, metric_config, base_rate, output_col, config)
}
\description{
Aplicar agravamento por risco
}
\keyword{internal}
