% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{apply_simulation_stage}
\alias{apply_simulation_stage}
\title{Apply a simulation stage}
\usage{
apply_simulation_stage(data, stage_config, approval_col, score_col, config)
}
\description{
Apply a simulation stage
}
\keyword{internal}
