% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{apply_suggested_cutoffs}
\alias{apply_suggested_cutoffs}
\title{Apply suggested cutoffs to data}
\usage{
apply_suggested_cutoffs(data, suggested_cutoffs, config)
}
\arguments{
\item{data}{Input data}

\item{suggested_cutoffs}{Named vector or list of suggested cutoffs}

\item{config}{Simulation configuration}
}
\value{
Data frame with applied cutoffs
}
\description{
Apply suggested cutoffs to data
}
