% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{apply_variable_based_stage}
\alias{apply_variable_based_stage}
\title{Apply variable-based simulation stage}
\usage{
apply_variable_based_stage(data, stage_config, approval_col, stage_col)
}
\description{
Apply variable-based simulation stage
}
\keyword{internal}
