% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{assign_simulated_defaults}
\alias{assign_simulated_defaults}
\title{Assign default outcomes for the final approved population}
\usage{
assign_simulated_defaults(
  data,
  policy,
  method = c("stochastic", "analytical"),
  quiet = FALSE
)
}
\value{
The data frame with a new \code{simulated_default} column.
}
\description{
Assign default outcomes for the final approved population
}
\keyword{internal}
