% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{calc_prob_aggravation}
\alias{calc_prob_aggravation}
\title{Calculate stressed default probability via aggravation factor}
\usage{
calc_prob_aggravation(data, policy, scenario_obj)
}
\description{
Calculate stressed default probability via aggravation factor
}
\keyword{internal}
