% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{calc_prob_monotonic}
\alias{calc_prob_monotonic}
\title{Calculate default probability for monotonic increase stress}
\usage{
calc_prob_monotonic(swap_ins, score_col, scenario)
}
\description{
Calculate default probability for monotonic increase stress
}
\keyword{internal}
