% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_decile_stats}
\alias{calculate_decile_stats}
\title{Calculate detailed decile statistics}
\usage{
calculate_decile_stats(data, score_col, default_col = "observed_default")
}
\arguments{
\item{data}{Input data}

\item{score_col}{Score column name}

\item{default_col}{Default column name}
}
\value{
Data frame with decile statistics
}
\description{
Calculate detailed decile statistics
}
