% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{calculate_historical_rates}
\alias{calculate_historical_rates}
\title{Calcular taxas históricas}
\usage{
calculate_historical_rates(data, config)
}
\description{
Calcular taxas históricas
}
\keyword{internal}
