% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{calculate_new_approval}
\alias{calculate_new_approval}
\title{Calculate approval by new score}
\usage{
calculate_new_approval(data, score_col, new_approval_col)
}
\description{
Calculate approval by new score
}
\keyword{internal}
