% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{calculate_performance_metrics}
\alias{calculate_performance_metrics}
\title{Calcular métricas de performance}
\usage{
calculate_performance_metrics(data, historical_rates, config)
}
\description{
Calcular métricas de performance
}
\keyword{internal}
