% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_required_steps}
\alias{calculate_required_steps}
\title{Calculate required cutoff steps based on desired precision}
\usage{
calculate_required_steps(score_range, desired_precision = 10, max_steps = 100)
}
\arguments{
\item{score_range}{Range of the score}

\item{desired_precision}{Desired precision in score units}

\item{max_steps}{Maximum number of steps to consider}
}
\value{
Number of steps needed
}
\description{
Calculate required cutoff steps based on desired precision
}
\examples{
# For a score that ranges from 300 to 850, with 10-point precision
calculate_required_steps(score_range = c(300, 850), desired_precision = 10)
}
