% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{classify_scenarios}
\alias{classify_scenarios}
\title{Classify applicants into transition scenarios}
\usage{
classify_scenarios(data, policy, new_approval_col)
}
\description{
Classify applicants into transition scenarios
}
\keyword{internal}
