% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{create_by_score_plot}
\alias{create_by_score_plot}
\title{Create by-score visualization}
\usage{
create_by_score_plot(analysis_results)
}
\description{
Create by-score visualization
}
\keyword{internal}
