% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{create_tradeoff_plot}
\alias{create_tradeoff_plot}
\title{Create trade-off scatter plot}
\usage{
create_tradeoff_plot(analysis_results)
}
\description{
Create trade-off scatter plot
}
\keyword{internal}
