% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-parallel.R
\name{.parallel_map}
\alias{.parallel_map}
\title{Internal wrapper for parallel map operations}
\usage{
.parallel_map(
  .x,
  .f,
  ...,
  .parallel = FALSE,
  .options = NULL,
  .progress = FALSE
)
}
\arguments{
\item{.x}{A list or vector.}

\item{.f}{A function.}

\item{...}{Additional arguments.}

\item{.parallel}{Logical.}

\item{.options}{furrr options (passed as a list or furrr_options object).}

\item{.progress}{Logical.}
}
\description{
Internal wrapper for parallel map operations
}
\keyword{internal}
