% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-parallel.R
\name{.parallel_map_dfr}
\alias{.parallel_map_dfr}
\title{Internal wrapper for parallel map_dfr operations}
\usage{
.parallel_map_dfr(
  .x,
  .f,
  ...,
  .parallel = FALSE,
  .options = NULL,
  .progress = FALSE
)
}
\description{
Internal wrapper for parallel map_dfr operations
}
\keyword{internal}
