% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-parallel.R
\name{.parallel_map_lgl}
\alias{.parallel_map_lgl}
\title{Internal wrapper for parallel map_lgl operations}
\usage{
.parallel_map_lgl(
  .x,
  .f,
  ...,
  .parallel = FALSE,
  .options = NULL,
  .progress = FALSE
)
}
\description{
Internal wrapper for parallel map_lgl operations
}
\keyword{internal}
