% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-parallel.R
\name{.parallel_pmap}
\alias{.parallel_pmap}
\title{Internal wrapper for parallel pmap operations}
\usage{
.parallel_pmap(
  .l,
  .f,
  ...,
  .parallel = FALSE,
  .options = NULL,
  .progress = FALSE
)
}
\description{
Internal wrapper for parallel pmap operations
}
\keyword{internal}
