% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-parallel.R
\name{.setup_parallel}
\alias{.setup_parallel}
\title{Internal helper for parallel processing setup}
\usage{
.setup_parallel(..., env = parent.frame())
}
\arguments{
\item{...}{Additional arguments passed to the function, potentially containing \code{parallel} and \code{n_workers}.}

\item{env}{Environment. The environment where on.exit should be registered (usually parent.frame()).}
}
\value{
A list with parallel (logical) indicating if it should proceed in parallel.
}
\description{
Internal helper for parallel processing setup
}
\keyword{internal}
