% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{evaluate_cutoff_combinations}
\alias{evaluate_cutoff_combinations}
\title{Evaluate all cutoff combinations}
\usage{
evaluate_cutoff_combinations(
  data,
  config,
  cutoff_ranges,
  target_default_rate,
  min_approval_rate,
  parallel = FALSE,
  n_cores = NULL
)
}
\description{
Evaluate all cutoff combinations
}
\keyword{internal}
