% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{evaluate_single_combination}
\alias{evaluate_single_combination}
\title{Evaluate a single cutoff combination}
\usage{
evaluate_single_combination(
  combo_id,
  cutoff_combinations,
  data,
  config,
  target_default_rate,
  min_approval_rate
)
}
\description{
Evaluate a single cutoff combination
}
\keyword{internal}
