% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_wrapper.R
\name{extract_recommendations}
\alias{extract_recommendations}
\title{Extract recommendations from optimization results}
\usage{
extract_recommendations(opt_results, strategy)
}
\description{
Extract recommendations from optimization results
}
\keyword{internal}
