% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{find_optimal_results}
\alias{find_optimal_results}
\title{Find optimal results based on constraints and tradeoff score}
\usage{
find_optimal_results(results, target_default_rate, min_approval_rate)
}
\description{
Find optimal results based on constraints and tradeoff score
}
\keyword{internal}
