% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{find_pareto_frontier}
\alias{find_pareto_frontier}
\title{Find Pareto frontier of non-dominated solutions}
\usage{
find_pareto_frontier(analysis_data)
}
\description{
Find Pareto frontier of non-dominated solutions
}
\keyword{internal}
