% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{generate_cutoff_ranges}
\alias{generate_cutoff_ranges}
\title{Generate cutoff ranges for all scores}
\usage{
generate_cutoff_ranges(data, score_columns, cutoff_steps)
}
\description{
Generate cutoff ranges for all scores
}
\keyword{internal}
