% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_cutoff_sequence}
\alias{generate_cutoff_sequence}
\title{Generate a sequence of cutoff values}
\usage{
generate_cutoff_sequence(
  min_cutoff = 300,
  max_cutoff = 850,
  n_points = 50,
  method = c("linear", "quantile", "optimal"),
  data = NULL,
  score_col = NULL
)
}
\arguments{
\item{min_cutoff}{Minimum cutoff value}

\item{max_cutoff}{Maximum cutoff value}

\item{n_points}{Number of points to generate}

\item{method}{Method for generating sequence ("linear", "quantile", "optimal")}

\item{data}{Optional data for quantile-based generation}

\item{score_col}{Optional score column for quantile-based generation}
}
\value{
Sequence of cutoff values
}
\description{
Generate a sequence of cutoff values
}
