% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{generate_realistic_score}
\alias{generate_realistic_score}
\title{Generate realistic credit score distribution}
\usage{
generate_realistic_score(n, mean = 700, sd = 80, skew = 0)
}
\description{
Generate realistic credit score distribution
}
\keyword{internal}
