% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R, R/simulation.R
\name{get_final_approval_col}
\alias{get_final_approval_col}
\title{Get final approval column name}
\usage{
get_final_approval_col(simulation_stages, score_col)

get_final_approval_col(simulation_stages, score_col)
}
\description{
Get final approval column name

Get final approval column name
}
\keyword{internal}
