% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{new_credit_policy}
\alias{new_credit_policy}
\title{Low-level constructor for the \code{credit_policy} class}
\usage{
new_credit_policy(
  applicant_id_col = character(),
  score_cols = character(),
  current_approval_col = character(),
  actual_default_col = character(),
  risk_level_col = NULL,
  simulation_stages = list(),
  stress_scenarios = list()
)
}
\description{
Low-level constructor for the \code{credit_policy} class
}
\keyword{internal}
