% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stages.R
\name{new_credit_policy_stage}
\alias{new_credit_policy_stage}
\title{Define a Simulation Stage}
\usage{
new_credit_policy_stage(name, type, ...)
}
\arguments{
\item{name}{A character string for the name of the stage (e.g., "credit", "fraud", "conversion").}

\item{...}{Specific parameters for the stage type.}
}
\value{
A \code{credit_policy_stage} object.
}
\description{
These functions create definitions for different types of stages in a
credit policy simulation funnel.
}
\keyword{internal}
