% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening.R
\name{predict.credit_risk_screening}
\alias{predict.credit_risk_screening}
\title{Predict Risk Segmentation on New Data}
\usage{
\method{predict}{credit_risk_screening}(object, newdata, variable, output_col = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{credit_risk_screening}.}

\item{newdata}{A data frame containing the same base risk and candidate columns.}

\item{variable}{Character. The name of the variable whose segmentation recipe should be applied.}

\item{output_col}{Optional. The name of the new column.}

\item{...}{Not used.}
}
\value{
The \code{newdata} data frame with an additional column of sub-segments.
}
\description{
\code{predict.credit_risk_screening()} applies the sub-segmentation logic discovered
during screening for a specific variable to a new dataset.
}
