% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{print.credit_policy}
\alias{print.credit_policy}
\title{Print method for credit_policy}
\usage{
\method{print}{credit_policy}(x, ...)
}
\arguments{
\item{x}{A \code{credit_policy} object.}

\item{...}{Additional arguments passed to \code{print}.}
}
\description{
Print method for credit_policy
}
