% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_wrapper.R
\name{print_recommendations}
\alias{print_recommendations}
\title{Print optimization recommendations}
\usage{
print_recommendations(recommendations, strategy)
}
\description{
Print optimization recommendations
}
\keyword{internal}
