% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_calculate_tier_metrics}
\alias{rcpp_calculate_tier_metrics}
\title{Fast calculation of metrics per tier for a candidate variable}
\usage{
rcpp_calculate_tier_metrics(x, groups, defaults, n_bins)
}
\arguments{
\item{x}{NumericVector candidate variable values}

\item{groups}{IntegerVector base risk groups (ratings)}

\item{defaults}{IntegerVector binary target (0/1)}

\item{n_bins}{int number of bins for discretization}
}
\value{
List of metrics per tier
}
\description{
Fast calculation of metrics per tier for a candidate variable
}
\examples{
\dontrun{
x <- runif(100)
g <- rep(1:2, each = 50)
d <- rbinom(100, 1, 0.1)
rcpp_calculate_tier_metrics(x, g, d, 2)
}
}
