% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_credit_process}
\alias{simulate_credit_process}
\title{Execute credit process simulation}
\usage{
simulate_credit_process(data, config, parallel = FALSE, n_cores = NULL)
}
\arguments{
\item{data}{Input data}

\item{config}{Simulation configuration}

\item{parallel}{Whether to use parallel processing}

\item{n_cores}{Number of cores for parallel processing}
}
\value{
Simulation results
}
\description{
Execute credit process simulation
}
