% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_stage.default}
\alias{simulate_stage.default}
\title{Default simulator for unknown stage types}
\usage{
\method{simulate_stage}{default}(data, stage, policy)
}
\description{
Default simulator for unknown stage types
}
\keyword{internal}
