% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_stage.stage_cutoff}
\alias{simulate_stage.stage_cutoff}
\title{Simulate a cutoff-based stage}
\usage{
\method{simulate_stage}{stage_cutoff}(data, stage, policy)
}
\description{
Simulate a cutoff-based stage
}
\keyword{internal}
