% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_stage.stage_rate}
\alias{simulate_stage.stage_rate}
\title{Simulate a rate-based stage (e.g., conversion)}
\usage{
\method{simulate_stage}{stage_rate}(data, stage, policy)
}
\description{
Simulate a rate-based stage (e.g., conversion)
}
\keyword{internal}
