% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stages.R
\name{stage_cutoff}
\alias{stage_cutoff}
\title{Define a Cutoff Stage}
\usage{
stage_cutoff(name, cutoffs, observed_outcome_col = NULL)
}
\arguments{
\item{name}{A character string for the name of the stage (e.g., "credit_check").}

\item{cutoffs}{A named list where names are score columns and values are
the cutoffs to be applied.}

\item{observed_outcome_col}{The column in the original data that contains the
observed outcome for this stage (0 or 1), if it exists. Used for \code{keep_in} analysis.}
}
\description{
\code{stage_cutoff()} defines a stage where approval is determined by one or more
score cutoffs. An applicant passes if their score is >= the cutoff for all
defined scores in this stage.
}
\details{
Define a Cutoff Stage
}
\examples{
# Defines a stage that requires a 'new_score' of at least 650
credit_stage <- stage_cutoff(name = "credit_check", cutoffs = list(new_score = 650))
}
