% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stages.R
\name{stage_rate}
\alias{stage_rate}
\title{Define a Rate Stage}
\usage{
stage_rate(
  name,
  base_rate,
  observed_outcome_col = NULL,
  stress_by_score = NULL
)
}
\arguments{
\item{name}{A character string for the name of the stage (e.g., "fraud_check").}

\item{base_rate}{The base rate for the simulation (e.g., 0.4 for a 40\% conversion rate).}

\item{observed_outcome_col}{The column in the original data that contains the
observed outcome for this stage (0 or 1), if it exists.}

\item{stress_by_score}{An optional named list to stress the \code{base_rate} based on a
score's value. Uses a monotonic increase assumption. See \code{stress_monotonic_increase()}.}
}
\description{
\code{stage_rate()} defines a stage where "approval" (e.g., conversion, acceptance)
is determined by a simulated rate. This is useful for stages that are not
score-based.
}
\details{
Define a Rate Stage
}
\examples{
# Defines a stage with a 95\% pass rate (e.g., for fraud checks)
fraud_stage <- stage_rate(name = "fraud_check", base_rate = 0.95)
}
