% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{validate_credit_policy}
\alias{validate_credit_policy}
\title{Validator for the \code{credit_policy} class}
\usage{
validate_credit_policy(policy)
}
\value{
The validated \code{credit_policy} object, invisibly.
}
\description{
Validator for the \code{credit_policy} class
}
\keyword{internal}
