% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{validate_optimization_inputs}
\alias{validate_optimization_inputs}
\title{Validate optimization inputs}
\usage{
validate_optimization_inputs(
  data,
  config,
  cutoff_steps,
  target_default_rate,
  min_approval_rate
)
}
\description{
Validate optimization inputs
}
\keyword{internal}
