% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{visualize_tradeoffs}
\alias{visualize_tradeoffs}
\title{Visualize trade-off analysis}
\usage{
visualize_tradeoffs(analysis_results, type = "tradeoff")
}
\arguments{
\item{analysis_results}{Trade-off analysis results}

\item{type}{Type of visualization ("tradeoff", "pareto", "by_score")}
}
\value{
ggplot object
}
\description{
Visualize trade-off analysis
}
