% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{calc_prob_aggravation}
\alias{calc_prob_aggravation}
\title{Calculate default probability based on aggravation}
\usage{
calc_prob_aggravation(data, policy, scenario)
}
\description{
Calculate default probability based on aggravation
}
\keyword{internal}
