% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{calc_prob_monotonic}
\alias{calc_prob_monotonic}
\title{Calculate probability based on a monotonic score-to-rate trend}
\usage{
calc_prob_monotonic(data, score_col, params)
}
\description{
Calculate probability based on a monotonic score-to-rate trend
}
\keyword{internal}
