% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{classify_scenarios}
\alias{classify_scenarios}
\title{Classify scenarios based on current and new approval decisions}
\usage{
classify_scenarios(data, policy, new_approval_col)
}
\description{
Classify scenarios based on current and new approval decisions
}
\keyword{internal}
