% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{classify_scenarios}
\alias{classify_scenarios}
\title{Classify scenarios}
\usage{
classify_scenarios(
  data,
  current_approval_col,
  final_approval_col,
  scenario_col
)
}
\description{
Classify scenarios
}
\keyword{internal}
