% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimate_computation_requirements}
\alias{estimate_computation_requirements}
\title{Estimate computation requirements for optimization}
\usage{
estimate_computation_requirements(config, cutoff_steps, sample_size = 10000)
}
\arguments{
\item{config}{Simulation configuration}

\item{cutoff_steps}{Number of cutoff steps per score}

\item{sample_size}{Sample size for estimation}
}
\value{
List with computation estimates
}
\description{
Estimate computation requirements for optimization
}
