% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_efficient_cutoff_sequence}
\alias{generate_efficient_cutoff_sequence}
\title{Generate efficient cutoff sequences based on score distribution}
\usage{
generate_efficient_cutoff_sequence(
  data,
  score_col,
  n_steps = 10,
  method = c("quantile", "linear", "optimal")
)
}
\arguments{
\item{data}{Input data}

\item{score_col}{Score column name}

\item{n_steps}{Number of steps}

\item{method}{Method for generating steps ("quantile", "linear", "optimal")}
}
\value{
Sequence of cutoff values
}
\description{
Generate efficient cutoff sequences based on score distribution
}
