% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot.credit_sim_results}
\alias{plot.credit_sim_results}
\alias{plot.creditools_simulation_from_data}
\title{Plot method for simulate_from_data() results}
\usage{
\method{plot}{credit_sim_results}(x, ...)

\method{plot}{creditools_simulation_from_data}(x, ...)
}
\arguments{
\item{x}{A \code{creditools_simulation_from_data} object returned by
\code{simulate_from_data()}.}

\item{...}{Unused, included for S3 method compatibility.}
}
\value{
A \code{ggplot} object (invisibly), after drawing the plot.
}
\description{
Plot method for simulate_from_data() results
}
