% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_stage}
\alias{simulate_stage}
\title{Generic function to simulate a single stage}
\usage{
simulate_stage(data, stage, policy)
}
\description{
Generic function to simulate a single stage
}
\keyword{internal}
