% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_stage.stage_filter}
\alias{simulate_stage.stage_filter}
\title{Simulate a hard-filter stage (Binary Rules)}
\usage{
\method{simulate_stage}{stage_filter}(data, stage, policy)
}
\description{
Simulate a hard-filter stage (Binary Rules)
}
\keyword{internal}
