% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_swap_in_defaults}
\alias{simulate_swap_in_defaults}
\title{Simulate default outcomes for swap-in applicants}
\usage{
simulate_swap_in_defaults(data, policy)
}
\description{
Simulate default outcomes for swap-in applicants
}
\keyword{internal}
