% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{summarize_results}
\alias{summarize_results}
\title{Summarize simulation results}
\usage{
summarize_results(results, by = NULL)
}
\arguments{
\item{results}{Simulation results}

\item{by}{Grouping variables}
}
\value{
Data frame with summary statistics
}
\description{
Summarize simulation results
}
