% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{validate_simulation_inputs}
\alias{validate_simulation_inputs}
\title{Validate inputs for a simulation run}
\usage{
validate_simulation_inputs(data, policy)
}
\description{
Validate inputs for a simulation run
}
\keyword{internal}
